requirements.txt
cryptography>=41.0.0
pynacl>=1.5.0
google-generativeai>=0.4.0
